feat: REjig all the grammars for shared rules and tokens, starting with stored procs
# Conflicts:
#	core/src/main/antlr4/com/databricks/labs/remorph/parsers/snowflake/SnowflakeParser.g4

# Conflicts:
#	core/pom.xml

+++ core/src/test/scala/com/databricks/labs/remorph/parsers/snowflake/SnowflakeExpressionBuilderSpec.scala
@@ -459,7 +459,7 @@ class SnowflakeExpressionBuilderSpec
       verify(literal).DATE()
       verify(literal).TIMESTAMP()
       verify(literal).string()
-      verify(literal).DECIMAL()
+      verify(literal).INT()
       verify(literal).FLOAT()
       verify(literal).REAL()
       verify(literal).trueFalse()
